Major refactor of the application: - Added user login and registration system with user details stored in local storage. - Users can create multiple quizzes, so I added a "Quiz view". Includes a separate view if the user has no quizzes yet. All quizzes have play, edit and delete options. - Cleaned up the remaining code to store the question order of a given quiz in the quiz object itself rather than a separate object. - Various small refactors: for example, created a "modal window" function for various modal window operations; created a "smooth fade" function for interface changes so this can be optionally set. Moved most event listeners to the top. - More linking: clicking the logo will now bring you to the Quiz view; there is also a "back to your quizzes" option after you've played a quiz.
In short, the "Quiz Maker" part of "Quiz maker maker maker" is now
virtually complete. Need to test for bugs (e.g. what if a user tries to
play a quiz with no questions?). I'm sure I can make the code much more
compact as well. This was a good exercise in how complex coding
interfaces can be without a framework like React.